requirements.txt
pyscard>=2.3.0
websockets>=12.0
pystray>=0.19.5
Pillow>=10.0.0
pycryptodome>=3.20.0
